{{cookiecutter.project_directory}}/extract/randomuser/requirements.txt
attrs==23.2.0
boto3==1.34.64
botocore==1.34.64
jmespath==1.0.1
jsonlines==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
randomuser==1.6
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
# pip install boto3 python-dotenv randomuser jsonlines && pip freeze > requirements.txt
